requirements.txt
transformers
langchain
gradio
langchain_community
torch
sentencepiece
seaborn
sqlalchemy
py2cypher
faiss-cpu
networkx